# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Golang backend for Food Ordering portfolio app.",
        "title": "Food Ordering Backend",
        "contact": {
            "name": "Vladlen Tereshchenko",
            "url": "https://github.com/VladlenT",
            "email": "vladlent.dev@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get all users. Requires admin rights.",
                "operationId": "user-get-all",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "0-based page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "amount of entries per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.DTOsWithPagination"
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
        "/users/logout": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Logout. Requires auth.",
                "operationId": "user-logout",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "401": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get info about current user. Requires auth.",
                "operationId": "user-info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ResponseDTO"
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
        "/users/signin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Sign in",
                "operationId": "user-login",
                "parameters": [
                    {
                        "description": "User login data",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.AuthDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ResponseDTO"
                        }
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/users/signup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create new user",
                "operationId": "user-create",
                "parameters": [
                    {
                        "description": "User info",
                        "name": "auth",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.AuthDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/user.ResponseDTO"
                        }
                    },
                    "409": {
                        "description": ""
                    },
                    "422": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
...
        "common.PaginationDTO": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "user.AuthDTO": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.DTOsWithPagination": {
            "type": "object",
            "properties": {
                "pagination": {
                    "$ref": "#/definitions/common.PaginationDTO"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.ResponseDTO"
                    }
                }
            }
        },
        "user.ResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_admin": {
                    "type": "boolean"
                }
            }
        }
    }
}